pyproject.toml
[tool.poetry]
name = "drfhomeworks"
version = "0.1.0"
description = ""
authors = ["CatritMe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
